+++ targets/export-elasticsearch/src/ingester/contributions/__tests__/generateContent.test.ts
@@ -43,7 +43,7 @@ describe("generateContent", () => {
     };
 
     const result: ContributionContent = await generateContent(
-      mockContributions,
+      mockContributions.find(({id}) => id === contribution.genericAnswerId),
       contribution
     );
 
@@ -59,7 +59,7 @@ describe("generateContent", () => {
     };
 
     const result: ContributionContent = await generateContent(
-      mockContributions,
+      mockContributions.find(({id}) => id === contribution.genericAnswerId),
       contribution
     );
 
@@ -77,7 +77,7 @@ describe("generateContent", () => {
     };
 
     await expect(
-      generateContent(mockContributions, contribution)
+      generateContent(mockContributions.find(({id}) => id === contribution.genericAnswerId), contribution)
     ).rejects.toThrowError(
       `Aucune contribution générique a été retrouvée pour la contribution [10 - 3239] (id générique non trouvé : unknown-type)`
     );
@@ -92,7 +92,7 @@ describe("generateContent", () => {
     };
 
     await expect(
-      generateContent(mockContributions, contribution)
+      generateContent(mockContributions.find(({id}) => id === contribution.genericAnswerId), contribution)
     ).rejects.toThrowError(
       "La contribution [2 - 1516] ne peut pas référencer une générique qui n'a pas de réponse"
     );
@@ -114,7 +114,7 @@ describe("generateContent", () => {
     (fetchFicheSp as jest.Mock).mockResolvedValue(mockFicheSpContent);
 
     const result: ContributionContent = await generateContent(
-      mockContributions,
+      mockContributions.find(({id}) => id === contribution.genericAnswerId),
       contribution
     );
 
@@ -142,7 +142,7 @@ describe("generateContent", () => {
     (fetchFicheSp as jest.Mock).mockResolvedValue(mockFicheSpContent);
 
     const result: ContributionContent = await generateContent(
-      mockContributions,
+      mockContributions.find(({id}) => id === contribution.genericAnswerId),
       contribution
     );
 
@@ -162,7 +162,7 @@ describe("generateContent", () => {
     };
 
     await expect(
-      generateContent(mockContributions, contribution)
+      generateContent(mockContributions.find(({id}) => id === contribution.genericAnswerId), contribution)
     ).rejects.toThrowError(
       'Type de contribution generic inconnu "unknown_type" for [GENERIC_UNKNOWN_TYPE]'
     );
@@ -175,7 +175,7 @@ describe("generateContent", () => {
     };
 
     const result: ContributionContent = await generateContent(
-      mockContributions,
+      mockContributions.find(({id}) => id === contribution.genericAnswerId),
       contribution
     );