+++ targets/hasura/migrations/default/1715781737558_alter_table_contribution_question_messages_add_column_content_agreement_without_legal/down.sql
@@ -1,0 +1,2 @@
+alter table "contribution"."question_messages" drop column "content_not_handled_without_legal";
+alter table "contribution"."question_messages" drop column "content_agreement_without_legal";